Yosys Synthesis Setup: Disable Convert Memory to Reg
Add -nomem2reg flag to Yosys read_verilog command.
read_verilog is the verilog frontend of Yosys. By default, it converts blockram structures
to registers if certain conditions are met. In our case, the blockrams from all benchmark
files are converted to registers, which crippeles the design. This is why we disable this
feature for all designs.

Signed-off-by: Benjamin Barzen <bbarzen@berkeley.edu>